requirements.txt
asgiref==3.7.2
cloudinary==1.39.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.25
django-autoslug==1.9.9
django-summernote==0.8.20.0
gunicorn==21.2.0
psycopg2==2.9.9
pytz==2024.1
slugify==0.0.1
sqlparse==0.4.4
